revise comment Model & controller

+++ controllers/comments.js
@@ -3,14 +3,15 @@ const appError = require('../helper/appError');
 const Comment = require('../model/comment');
 
 const CommentController = {
-  async getComments(req, res) {
-    const comments = await Comment.find().populate({
+  async getComments(req, res, next) {
+    const { articleID } = req.params;
+    const comments = await Comment.find({ articleID }).populate({
       path: 'author',
       select: '_id name avatar',
     });
     successHandler(res, comments);
   },
-  async createComments(req, res) {
+  async createComments(req, res, next) {
     const { articleID, author, comment } = req.body;
     if (articleID && author && comment) {
       await Comment.create({
@@ -27,12 +28,17 @@ const CommentController = {
       );
     }
   },
-  async deleteAllComments(req, res) {
-    const comments = await Comment.deleteMany({});
-    successHandler(res, comments);
-  },
-  async deleteComments(req, res) {
+  async deleteComments(req, res, next) {
     const { id } = req.params;
+    const comment = await Comment.findById(id);
+    // 非本人不能刪除評論
+    if (req.user._id !== comment.author._id) {
+      return appError(
+        401,
+        'Bad Request Error - You do not have permission to delete this comment',
+        next
+      );
+    }
     await Comment.findByIdAndDelete(id)
       .then((result) => {
         if (!result) {
@@ -41,10 +47,36 @@ const CommentController = {
         CommentController.getComments(req, res);
       })
       .catch(() => appError(400, 'Bad Request Error - ID not found', next));
+
+    // await Comment.findById(id)
+    //   .then((result) => {
+    //     if (!result) {
+    //       return appError(400, 'Bad Request Error - Failed to get data', next);
+    //     }
+    //     // 非本人不能刪除評論
+    //     if(req.user._id !== result.author._id) {
+    //       return appError(401, 'Bad Request Error - You do not have permission to delete this comment', next);
+    //     }
+    //     await Comment.findByIdAndDelete(id)
+    //     .then(()=>{
+    //       CommentController.getComments(req, res);
+    //     })
+    //     .catch(() => appError(400, 'Bad Request Error - ID not found', next));
+    //   })
+    //   .catch(() => appError(400, 'Bad Request Error - ID not found', next));
   },
-  async editComments(req, res) {
+  async editComments(req, res, next) {
     const { body } = req;
     const { id } = req.params;
+    const comment = await Comment.findById(id);
+    // 非本人不能修改評論
+    if (req.user._id !== comment.author._id) {
+      return appError(
+        401,
+        'Bad Request Error - You do not have permission to delete this comment',
+        next
+      );
+    }
     await Comment.findByIdAndUpdate(id, body)
       .then((result) => {
         if (!result) {
@@ -53,6 +85,10 @@ const CommentController = {
         CommentController.getComments(req, res);
       })
       .catch(() => appError(400, 'Bad Request Error - ID not found', next));
+  },
+  async deleteAllComments(req, res, next) {
+    const comments = await Comment.deleteMany({});
+    successHandler(res, comments);
   },
 };
